Fix bug with Transpose codec.
This commit fixes a bug where subsequent codecs were not
aware of the Transpose codec's encoded representation. Meaning if
the transpose changed the encoded representation's shape compared to
its decoded representation, this information was not propagated to
subsequent codecs when decoding chunks, which lead to an
Out of index exception getting thrown when decoding a chunk
of an array that contains the Transpose codec in its codec chain.