Enforce shard index codec chain invariant.
The V3 spec insists that the codec chain for a shard index array
must not include compression codecs like Gzip since it would make
it impractical to calculate index bytes size. This commit
ensures this invariant is not violated by making a shard's index
codec chain correct by construction. That is, specifying a non-fixed
size bytes-to-bytes codec in the `index_codec` field is a compile time
error.