Improve memory usage when reading array from store.
This ensures that a chunk is visited and decoded once without
also storing it's decoded contents in a hashtable, which leads
to unnecessary usage in memory due to storing temporary chunk
array data. This commit improves on this by getting all elements
from a chunk at once, but ensures that the slice's coordinate C-order
position in the requested slice is preserved.