Refactor to add support for asynchronous read/write.
This commit does a big refactor to add support for concurrent I/O
via Lwt (or any other concurrency library we may want to support in
the future). It does so by functorizing all the code that does I/O and
generalizes it such that it can work for Lwt and for synchronous
reads and writes.