Minimize `get_partial` & `set_partial` calls.
calling these for every requested inner chunk of a shard is
expensive can can lead to a `Unix.EMFILE` error. This commit
attempts to minimize the number of calls by making sure these are
called once on all requested inner chunks of a shard. This should
lead to a significant decrease in the number of open files open
for read/write at the same time when using a Zarr_lwt `FilesystemStore`.