Enforce correct sequencing of codecs in codec chain.
This ensures that creation of a codec chain fails if the sequencing of
codecs in the list is not correct. That is if the sequencing is not
exactly: arraytoarray list -> 1 array-to-bytes -> bytestobytes list.
That is one must supply the codecs in the order in which they intend to
apply the encode/decode sequence.