Add `ZipStore` storage backend.
This commit adds a ZipStore storage backend as described in the
specification https://github.com/zarr-developers/zarr-specs/pull/311 .
Note that the implementation loads the entire zip archive into memory so
care must be taken to ensure the zip archive is not too big to fit into
the machine's memory. To use a ZipStore impelementation that does not
load the archive into memory see `examples/zipstore.ml`.