Use `camlzip` package to implement a `ZipStore`.
After recently reading the licensing terms `camlzip`, I realized that
although it is licensed under LGPL, it contains a special exception that
allows library authors to link to camlzip without being bound to the
licensing terms of LGPL. The exception reads:

    As a special exception to the GNU Lesser General Public License, you
    may link, statically or dynamically, a "work that uses the Library" with
    a publicly distributed version of the Library to produce an executable
    file containing portions of the Library, and distribute that executable
    file under terms of your choice, without any of the additional requirements
    listed in clause 6 of the GNU Lesser General Public License.

I was initially not aware of this exception in the license and that
drove the decision to implement ZipStore using `zipc` (which reads an
entire zip archive into memory). This commit replaces `zipc` with
`camlzip` and as a result re-works the `ZipStore` implementation.